requirements.txt
flask>=3.0.0
python-escpos>=3.1
Pillow>=10.0.0
requests>=2.31.0

# Raspberry Pi / hardware — comment out for local dev
displayhatmini>=0.0.2
RPi.GPIO>=0.7.1
